+++ Acme Handy Worker/src/main/java/services/PersonalRecordService.java
@@ -10,7 +10,10 @@ import org.springframework.util.Assert;
 import repositories.PersonalRecordRepository;
 import security.LoginService;
 import security.UserAccount;
+import domain.Curricula;
+import domain.HandyWorker;
 import domain.PersonalRecord;
+import domain.ProfessionalRecord;
 
 
 @Service
@@ -23,8 +26,11 @@ public class PersonalRecordService {
 	
 	//Supporting Services -----
 	
-	//@Autowired
-	//private SomeService serviceName 
+	@Autowired
+	private CurriculaService curriculaService; 
+	
+	@Autowired
+	private ActorService actorService;
 	
 	//Constructors -----
 	public PersonalRecordService(){
@@ -33,8 +39,7 @@ public class PersonalRecordService {
 	
 	//Simple CRUD methods -----
 	public PersonalRecord create(){
-		//Metodo general para todas los servicios, es probable 
-		//que sea necesario añadir atributos consistentes con la entity.
+
 		PersonalRecord res = new PersonalRecord();
 		return res;
 	}
@@ -48,27 +53,57 @@ public class PersonalRecordService {
 	}
 	
 	public PersonalRecord save(PersonalRecord a){
-		//puede necesitarse control de versiones por concurrencia del objeto.
-		//puede necesitarse comprobar que el usuario que va a guardar el objeto es el dueño
-		Assert.isTrue(true);//modificar para condiciones especificas
 		
-		UserAccount userAccount = LoginService.getPrincipal();
-		// modificar para aplicarlo a la entidad correspondiente.
-		//Assert.isTrue(a.getUserAccount().equals(userAccount));
+		//si el HandyWorker tiene una curricula se le guarda/actualiza el PR, si no simplemente se guarda PR sin vincular.
 		
-		personalRecordRepository.save(a);
-		return a;
+		PersonalRecord res = null;
+		boolean isInCurricula = false;
+		boolean hasCurricula = false;
+		Assert.isTrue(LoginService.hasRole("HANDYWORKER"));
+		UserAccount logged = LoginService.getPrincipal();
+		
+		for (Curricula c : curriculaService.findAll()) {
+			if(c.getPersonalRecord().equals(a)){//vemos si esta dentro de alguna curricula
+				if(c.getHandyWorker().getUserAccount().equals(logged)){//y que el dueño es el logueado
+					res = personalRecordRepository.saveAndFlush(a);//actualizamos la curricula
+				}
+					isInCurricula=true;
+					//System.out.println("Se ha encontrado una curricula que contiene al personalRecord, se actualiza ese record.");
+			}
+		}
+		if (!isInCurricula){// si no tiene curricula asignada
+			for (Curricula cu : curriculaService.findAll()) {
+				if(cu.getHandyWorker().getUserAccount().equals(logged)){
+					res = personalRecordRepository.saveAndFlush(a);
+					cu.setPersonalRecord(res);
+					curriculaService.save(cu);
+					hasCurricula = true;
+					//System.out.println("se ha encontrado una curricula para el usuario logueado, se le asigna el personalrecord");
+				}
+			}
+		}
+			if(!hasCurricula){
+				res = personalRecordRepository.saveAndFlush(a);
+				//System.out.println("no existe una curricula para el handyworker logueado, se guarda el personal record sin asignar.");
+			}
+		Assert.notNull(res);
+		return res;
 	}
 	
 	public void delete(PersonalRecord a){
-		//puede necesitarse comprobar que el usuario que va a guardar el objeto es el dueño
-		Assert.isTrue(true);//modificar para condiciones especificas.(data constraint)
 		
-		UserAccount userAccount = LoginService.getPrincipal();
-		// modificar para aplicarlo a la entidad correspondiente.
-		//Assert.isTrue(a.getUserAccount().equals(userAccount));
-		
-		personalRecordRepository.delete(a);
+		Assert.isTrue(LoginService.hasRole("HANDYWORKER"));
+		boolean hasCurricula = false;
+		//vemos si existe la curricula, si, si llamamos al delete de curricula si no borramos el personalrecord
+		for (Curricula c : curriculaService.findAll()) {
+			if(c.getPersonalRecord().equals(a)){
+				curriculaService.delete(c);
+				hasCurricula = true;
+			}
+		}
+		if(!hasCurricula){
+			personalRecordRepository.delete(a);
+		}
 	}
 	
 	//Other business methods -----
